requirements.txt
# Core requirements
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
uuid==1.30
